First draft: markdown-parser into its own proper component
