+++ client/ListMapper/list-mapper/src/resources/common/elements/markdown-parser.html
@@ -18,12 +18,6 @@
     >
     </radial-context-menu>
 
-                      <button 
-                        class="btn btn-outline-secondary btn-sm"
-                        click.delegate="addCodeBlock()"
-                      >Add new .md Parser
-                      </button>
-
                       <segmented-button
                         button-text.bind="segmentedButtonText"
                       >
@@ -37,7 +31,6 @@
     <div 
       class="markdown-parser-result-container" 
       enable-drag="allow-from.bind: DRAG_BUTTON_CLASS;"
-      repeat.for="codeBlock of codeBlocks"
     >
       <more-options-menu>
         <button 
@@ -73,7 +66,7 @@
         >
         <!-- This is unfortunately an ugly way of writing code. -->
         <!-- Reason is pre preserves white space. EVERY white space -->
-        <pre><code class="language-js" ref="insertCodeRef"><div 
+        <pre><code class="language-js" ref="insertCodeRef"><div
           repeat.for="line of splittedLines"
           innerhtml.bind="line"
           class="line-number"
